New methods get/delete and cleanup

+++ server/src/main/java/org/aramuk/glist/TableHandler.java
@@ -1,5 +1,6 @@
 package org.aramuk.glist;
 
+import java.net.HttpURLConnection;
 import java.util.Arrays;
 import java.util.List;
 
@@ -44,12 +45,12 @@ import com.amazonaws.services.dynamodbv2.model.*;
  */
 public class TableHandler {
 
-    public static final String DEFAULT_TABLE_NAME = "items";
-    public static final String ATTR_DEVICE_ID = "device_id";
-    public static final String ATTR_PARENT_ID = "parent_item_id";
-    public static final String ATTR_ITEM_INFO = "item_info";
-    public static final String ATTR_ITEM_ID = "item_id";
-    public static final String ATTR_ITEM_VALUE = "value";
+    private static final String DEFAULT_TABLE_NAME = "items";
+    private static final String ATTR_DEVICE_ID = "device_id";
+    private static final String ATTR_PARENT_ID = "parent_item_id";
+    private static final String ATTR_ITEM_INFO = "item_info";
+    static final String ATTR_ITEM_ID = "item_id";
+    static final String ATTR_ITEM_VALUE = "value";
 
     private AmazonDynamoDB dynamoDBClient;
     private String tableName;
@@ -59,7 +60,7 @@ public class TableHandler {
             if (args[0].equalsIgnoreCase("init")) {
                 TableHandler handler = new TableHandler();
                 if (handler.checkTableExists()) {
-                    handler.deleteTable();
+                    handler.dropTableInternal();
                 }
                 handler.createTable();
                 handler.releaseResources();
@@ -99,6 +100,11 @@ public class TableHandler {
             return new DynamoDB(dynamoDBClient);
         }
         throw new IllegalStateException("DynamoDBClient not initialized. Using object after releasing resources?");
+    }
+
+    private Table getTable() {
+        DynamoDB dynamoDB = getDynamoDB();
+        return dynamoDB.getTable(tableName);
     }
 
     /**
@@ -142,15 +148,23 @@ public class TableHandler {
         }
     }
 
-    public void deleteTable() {
-        DynamoDB dynamoDB = getDynamoDB();
-        Table table = dynamoDB.getTable(tableName);
+    public void dropTable() {
+        if (tableName.equals(DEFAULT_TABLE_NAME)) {
+            throw new IllegalStateException("Can not drop default table");
+        }
+        dropTableInternal();
+    }
+
+    private void dropTableInternal() {
+        if (!checkTableExists()) {
+            throw new IllegalStateException("Table " + tableName + " does not exist");
+        }
+        Table table = getTable();
         try {
             System.out.println("Issuing DeleteTable request for " + tableName);
             table.delete();
             System.out.println("Waiting for " + tableName
                     + " to be deleted...this may take a while...");
-
             table.waitForDelete();
         } catch (Exception e) {
             System.err.println("DeleteTable request failed for " + tableName);
@@ -166,13 +180,30 @@ public class TableHandler {
      * @param values List containing Map of entries
      */
     public void addItem(String deviceId, String parentId, List values) {
-        DynamoDB dynamoDB = getDynamoDB();
-        Table table = dynamoDB.getTable(tableName);
+        // ToDo: Add code to not replace existing entries
+        Table table = getTable();
         Item item = new Item()
                 .withPrimaryKey(ATTR_DEVICE_ID, deviceId, ATTR_PARENT_ID, parentId)
                 .withList(ATTR_ITEM_INFO, values);
         PutItemOutcome outcome = table.putItem(item);
-        System.out.println("AddItemDone with HTTP Status code: "
-                + outcome.getPutItemResult().getSdkHttpMetadata().getHttpStatusCode());
+        int httpStatusCode = outcome.getPutItemResult().getSdkHttpMetadata().getHttpStatusCode();
+        if (httpStatusCode != HttpURLConnection.HTTP_OK) {
+            throw new RuntimeException("DynamoDB addItem failed with http status code " + httpStatusCode);
+        }
+    }
+
+    public Item getItem(String deviceId, String parentId) {
+        Table table = getTable();
+        return table.getItem(ATTR_DEVICE_ID, deviceId, ATTR_PARENT_ID, parentId);
+    }
+
+    public void deleteItem(String deviceId, String parentId) {
+        // ToDo: Add code to check that there are no child items
+        Table table = getTable();
+        DeleteItemOutcome outcome = table.deleteItem(ATTR_DEVICE_ID, deviceId, ATTR_PARENT_ID, parentId);
+        int httpStatusCode = outcome.getDeleteItemResult().getSdkHttpMetadata().getHttpStatusCode();
+        if (httpStatusCode != HttpURLConnection.HTTP_OK) {
+            throw new RuntimeException("DynamoDB deleteItem failed with http status code " + httpStatusCode);
+        }
     }
 }
